Add RemotingClient_MBRO_Lazy demo
The `RemotingClient_MBRO_Lazy` demo is similar to the `RemotingClient_MBRO`
demo and causes the creation of a `MarshalByRefObject` instance on the
server side, which gets returned as `ObjRef` for remote interaction.
However, instead of requiring the deserialization of a `DataTable` object
to parse XAML, it utilizes `System.Lazy<T>`, which creates an instance
of the given type parameter `T` on serialization, which is retrievable
by the client in the server's response.